pyproject.toml
[tool.poetry]
name = "keyclock-auth-flow"
version = "0.1.0"
description = ""
authors = ["Abdullah Alqahtani <0xqahta0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic = "^2.10.4"
email-validator = "^2.2.0"
jwt = "^1.3.1"
pyjwt = "^2.10.1"
faker = "^33.1.0"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"